pyproject.toml
[tool.poetry]
name = "ai-image-hackaton"
version = "0.1.0"
description = ""
authors = ["canaparro <marcelocanaparro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.47.1"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
opensearch-py = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
